# docs/openapi.yaml
openapi: 3.0.3
info:
  title: MediaPulse API
  version: 1.0.0-beta.98
  description: API surface for MediaPulse. This document covers the published HTTP surface, while the markdown files in `docs/` remain the canonical source for operational notes and endpoint semantics.
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/movies/recent:
    get:
      summary: Recent watched movies
      tags: [Movies]
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
        - name: cursor
          in: query
          required: false
          schema:
            type: string
            nullable: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MoviesRecentResponse"
  /api/movies/{movieId}:
    get:
      summary: Movie details
      tags: [Movies]
      parameters:
        - name: movieId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MovieDetailsResponse"
  /api/movies/slug/{slug}:
    get:
      summary: Movie details by slug
      tags: [Movies]
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MovieDetailsResponse"
  /api/movies/library:
    get:
      summary: Movies library
      tags: [Movies]
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
        - name: cursor
          in: query
          required: false
          schema:
            type: string
            nullable: true
        - name: unwatched
          in: query
          required: false
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MoviesLibraryResponse"
...
  /api/movies/companies/{slug}:
    get:
      summary: Movie company details by slug
      tags: [Movies]
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MovieCompanyDetailsResponse"
  /api/movies/lists:
    get:
      summary: List manual movie lists
      tags: [Movies]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MovieListSummaryDto"
    post:
      summary: Create manual movie list
      tags: [Movies]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MovieListCreateRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MovieListSummaryDto"
  /api/movies/lists/{slug}:
    get:
      summary: Manual movie list details
      tags: [Movies]
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MovieListDetailsResponse"
  /api/movies/lists/{listId}/order:
    post:
      summary: Update manual movie list order
      tags: [Movies]
      parameters:
        - name: listId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MovieListOrderUpdateRequest"
      responses:
        "200":
          description: OK
  /api/movies/lists/{listId}/cover:
    patch:
      summary: Update manual movie list cover
      tags: [Movies]
      parameters:
        - name: listId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MovieListCoverUpdateRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MovieListSummaryDto"
  /api/movies/{movieId}/lists:
    post:
      summary: Attach movie to manual list
      tags: [Movies]
      parameters:
        - name: movieId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MovieListAttachRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MovieListSummaryDto"
  /api/movies/{movieId}/lists/{listId}:
    delete:
      summary: Remove movie from manual list
      tags: [Movies]
      parameters:
        - name: movieId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: listId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
...
  /api/movies/terms/{kind}/{slug}:
    get:
      summary: Movie term details by kind and slug
      tags: [Movies]
      parameters:
        - name: kind
          in: path
          required: true
          schema:
            type: string
            enum: [genre, tag, GENRE, TAG]
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MovieTermDetailsResponse"
  /api/movies/terms/search:
    get:
      summary: Search local movie terms
      tags: [Movies]
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: kind
          in: query
          required: true
          schema:
            type: string
            enum: [genre, tag, GENRE, TAG]
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 8
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MovieTermSuggestionDto"
...
  /api/movies/year/{year}:
    get:
      summary: Movies by year
      tags: [Movies]
      parameters:
        - name: year
          in: path
          required: true
          schema:
            type: integer
        - name: limitWatched
          in: query
          required: false
          schema:
            type: integer
            default: 200
        - name: limitUnwatched
          in: query
          required: false
          schema:
            type: integer
            default: 200
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MoviesByYearResponse"
...
  /api/movies/{movieId}/credits/tmdb-candidates:
    post:
      summary: Fetch extra TMDb credit candidates for a movie
      tags: [Movies]
      parameters:
        - name: movieId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MovieTmdbCreditCandidatesResponse"
  /api/movies/{movieId}/credits/from-tmdb:
    post:
      summary: Import a specific TMDb credit into a movie
      tags: [Movies]
      parameters:
        - name: movieId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MovieTmdbCreditImportRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PersonCreditDto"
  /api/movies/{movieId}/people:
    post:
      summary: Link existing local person to movie
      tags: [Movies]
      parameters:
        - name: movieId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PersonLinkRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PersonCreditDto"
  /api/movies/{movieId}/people/{personId}:
    delete:
      summary: Remove a movie credit category and freeze manual curation
      tags: [Movies]
      parameters:
        - { name: movieId, in: path, required: true, schema: { type: integer, format: int64 } }
        - { name: personId, in: path, required: true, schema: { type: integer, format: int64 } }
        - { name: category, in: query, required: true, schema: { type: string, enum: [CAST, DIRECTING, WRITING] } }
      responses:
        "200": { description: Removed }
  /api/movies/{movieId}/terms:
    post:
      summary: Create manual movie term
      tags: [Movies]
      parameters:
        - name: movieId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MovieTermCreateRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MovieTermDto"
  /api/movies/{movieId}/terms/{termId}/visibility:
    post:
      summary: Update movie-local term visibility
      tags: [Movies]
      parameters:
        - name: movieId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: termId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MovieTermVisibilityRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MovieTermDto"
  /api/movies/terms/{termId}/visibility:
    post:
      summary: Update global term visibility
      tags: [Movies]
      parameters:
        - name: termId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MovieTermVisibilityRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MovieTermDto"
...
  /api/movies/{movieId}/watches:
    post:
      summary: Create manual watch for existing movie
      tags: [Movies]
      parameters:
        - name: movieId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExistingMovieWatchCreateRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManualMovieWatchCreateResponse"
  /api/movies/{movieId}/watches/{watchId}:
    delete:
      summary: Delete movie watch
      tags: [Movies]
      parameters:
        - name: movieId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: watchId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
  /api/movies/search:
    get:
      summary: Search movies by title or slug
      tags: [Movies]
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MoviesSearchResponse"
  /api/movies/summary:
    get:
      summary: Movies summary
      tags: [Movies]
      parameters:
        - name: range
          in: query
          required: false
          schema:
            type: string
            enum: [month, year, custom]
            default: month
        - name: start
          in: query
          required: false
          schema:
            type: string
            format: date-time
        - name: end
          in: query
          required: false
          schema:
            type: string
            format: date-time
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MoviesSummaryResponse"
...
components:
...
  schemas:
...
    MovieTermVisibilityRequest:
      type: object
      properties:
        hidden:
          type: boolean
      required: [hidden]
...
    MovieListSummaryDto:
      type: object
      properties:
        listId:
          type: integer
          format: int64
        name:
          type: string
        slug:
          type: string
        description:
          type: string
          nullable: true
        itemCount:
          type: integer
          format: int64
        coverMovieId:
          type: integer
          format: int64
          nullable: true
        coverUrl:
          type: string
          nullable: true
        previewMovies:
          type: array
          items:
            $ref: "#/components/schemas/MovieListPreviewMovieDto"
      required: [listId, name, slug, itemCount, previewMovies]
    MovieListDetailsResponse:
      type: object
      properties:
        listId:
          type: integer
          format: int64
        name:
          type: string
        slug:
          type: string
        description:
          type: string
          nullable: true
        coverMovieId:
          type: integer
          format: int64
          nullable: true
        coverUrl:
          type: string
          nullable: true
        movieCount:
          type: integer
          format: int64
        watchedMoviesCount:
          type: integer
          format: int64
        movies:
          type: array
          items:
            $ref: "#/components/schemas/MovieLibraryCardDto"
      required: [listId, name, slug, movieCount, watchedMoviesCount, movies]
    MovieListCreateRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
      required: [name]
    MovieListAttachRequest:
      type: object
      properties:
        listId:
          type: integer
          format: int64
          nullable: true
        name:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
    MovieListOrderUpdateRequest:
      type: object
      properties:
        movieIds:
          type: array
          items:
            type: integer
            format: int64
      required: [movieIds]
    MovieListCoverUpdateRequest:
      type: object
      properties:
        coverMovieId:
          type: integer
          format: int64
          nullable: true
    MovieDetailsResponse:
      type: object
      properties:
        movieId:
          type: integer
          format: int64
        title:
          type: string
        originalTitle:
          type: string
        slug:
          type: string
          nullable: true
        year:
          type: integer
          nullable: true
        description:
          type: string
          nullable: true
        coverUrl:
          type: string
          nullable: true
        images:
          type: array
          items:
            $ref: "#/components/schemas/MovieImageDto"
        watches:
          type: array
          items:
            $ref: "#/components/schemas/MovieWatchDto"
        externalIds:
          type: array
          items:
            $ref: "#/components/schemas/MovieExternalIdDto"
        lists:
          type: array
          items:
            $ref: "#/components/schemas/MovieListSummaryDto"
        companies:
          type: array
          items:
            $ref: "#/components/schemas/MovieCompanyDto"
        people:
          type: array
          items:
            $ref: "#/components/schemas/PersonCreditDto"
        creditsCurated:
          type: boolean
        terms:
          type: array
          items:
            $ref: "#/components/schemas/MovieTermDto"
        collection:
          allOf:
            - $ref: "#/components/schemas/MovieCollectionDto"
          nullable: true
        rating:
          allOf:
            - $ref: "#/components/schemas/MediaRatingDto"
          nullable: true
        comments:
          type: array
          items:
            $ref: "#/components/schemas/MediaCommentDto"
        enrichment:
          $ref: "#/components/schemas/MovieAutomaticEnrichmentDto"
      required:
        [
          movieId,
          title,
          originalTitle,
          images,
          watches,
          externalIds,
          lists,
          companies,
          people,
          creditsCurated,
          terms,
          comments,
          enrichment,
        ]
...
    MovieCompanyDetailsResponse:
      type: object
      properties:
        companyId:
          type: integer
          format: int64
        tmdbId:
          type: string
        name:
          type: string
        slug:
          type: string
        logoUrl:
          type: string
          nullable: true
        originCountry:
          type: string
          nullable: true
        companyType:
          $ref: "#/components/schemas/MovieCompanyTypeDto"
        movieCount:
          type: integer
          format: int64
        watchedMoviesCount:
          type: integer
          format: int64
        movies:
          type: array
          items:
            $ref: "#/components/schemas/MovieLibraryCardDto"
      required:
        [
          companyId,
          tmdbId,
          name,
          slug,
          companyType,
          movieCount,
          watchedMoviesCount,
          movies,
        ]
    MovieTermSuggestionDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        slug:
          type: string
        kind:
          $ref: "#/components/schemas/MovieTermKindDto"
        source:
          $ref: "#/components/schemas/MovieTermSourceDto"
        hiddenGlobally:
          type: boolean
      required: [id, name, slug, kind, source, hiddenGlobally]
    MovieTermDetailsResponse:
      type: object
      properties:
        termId:
          type: integer
          format: int64
        name:
          type: string
        slug:
          type: string
        kind:
          $ref: "#/components/schemas/MovieTermKindDto"
        source:
          $ref: "#/components/schemas/MovieTermSourceDto"
        movieCount:
          type: integer
          format: int64
        watchedMoviesCount:
          type: integer
          format: int64
        movies:
          type: array
          items:
            $ref: "#/components/schemas/MovieLibraryCardDto"
      required:
        [
          termId,
          name,
          slug,
          kind,
          source,
          movieCount,
          watchedMoviesCount,
          movies,
        ]
...
    MovieTmdbCreditCandidatesResponse:
      type: object
      properties:
        movieId:
          type: integer
          format: int64
        reconciledCount:
          type: integer
        candidateCount:
          type: integer
        groups:
          type: array
          items:
            $ref: "#/components/schemas/MovieTmdbCreditCandidateGroupDto"
      required: [movieId, reconciledCount, candidateCount, groups]
    MovieTmdbCreditImportRequest:
      type: object
      properties:
        personTmdbId:
          type: string
        creditType:
          $ref: "#/components/schemas/MovieCreditTypeDto"
        department:
          type: string
          nullable: true
        job:
          type: string
          nullable: true
        characterName:
          type: string
          nullable: true
        billingOrder:
          type: integer
          nullable: true
      required: [personTmdbId, creditType]
    MoviesRecentResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/MovieCardDto"
        nextCursor:
          type: string
          nullable: true
      required: [items]
...
    MoviesLibraryResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/MovieLibraryCardDto"
        nextCursor:
          type: string
          nullable: true
      required: [items]
...
    MoviesByYearResponse:
      type: object
      properties:
        year:
          type: integer
        range:
          $ref: "#/components/schemas/MoviesRangeDto"
        stats:
          $ref: "#/components/schemas/MoviesByYearStatsDto"
        watched:
          type: array
          items:
            $ref: "#/components/schemas/MovieYearWatchedDto"
        unwatched:
          type: array
          items:
            $ref: "#/components/schemas/MovieYearUnwatchedDto"
      required: [year, range, stats, watched, unwatched]
...