# server/src/main/kotlin/dev/marcal/mediapulse/server/controller/movies/MoviesController.kt
package dev.marcal.mediapulse.server.controller.movies

import dev.marcal.mediapulse.server.api.movies.ExistingMovieWatchCreateRequest
import dev.marcal.mediapulse.server.api.movies.ManualMovieWatchCreateResponse
import dev.marcal.mediapulse.server.api.movies.MovieCompanyDetailsResponse
import dev.marcal.mediapulse.server.api.movies.MovieDetailsResponse
import dev.marcal.mediapulse.server.api.movies.MovieListAttachRequest
import dev.marcal.mediapulse.server.api.movies.MovieListCoverUpdateRequest
import dev.marcal.mediapulse.server.api.movies.MovieListCreateRequest
import dev.marcal.mediapulse.server.api.movies.MovieListDetailsResponse
import dev.marcal.mediapulse.server.api.movies.MovieListOrderUpdateRequest
import dev.marcal.mediapulse.server.api.movies.MovieListSummaryDto
import dev.marcal.mediapulse.server.api.movies.MovieTermCreateRequest
import dev.marcal.mediapulse.server.api.movies.MovieTermDetailsResponse
import dev.marcal.mediapulse.server.api.movies.MovieTermDto
import dev.marcal.mediapulse.server.api.movies.MovieTermSuggestionDto
import dev.marcal.mediapulse.server.api.movies.MovieTermVisibilityRequest
import dev.marcal.mediapulse.server.api.movies.MovieTmdbCreditCandidatesResponse
import dev.marcal.mediapulse.server.api.movies.MovieTmdbCreditImportRequest
import dev.marcal.mediapulse.server.api.movies.MoviesByYearResponse
import dev.marcal.mediapulse.server.api.movies.MoviesLibraryResponse
import dev.marcal.mediapulse.server.api.movies.MoviesRecentResponse
import dev.marcal.mediapulse.server.api.movies.MoviesSearchResponse
import dev.marcal.mediapulse.server.api.movies.MoviesStatsResponse
import dev.marcal.mediapulse.server.api.movies.MoviesSummaryResponse
import dev.marcal.mediapulse.server.api.movies.PersonCreditDto
import dev.marcal.mediapulse.server.api.movies.PersonLinkRequest
import dev.marcal.mediapulse.server.repository.MovieQueryRepository
import dev.marcal.mediapulse.server.service.movie.ExistingMovieWatchCreateFlowService
import dev.marcal.mediapulse.server.service.movie.MovieCreditsService
import dev.marcal.mediapulse.server.service.movie.MovieListsService
import dev.marcal.mediapulse.server.service.movie.MovieTermsService
import dev.marcal.mediapulse.server.service.movie.MovieWatchRemovalService
import org.springframework.http.HttpStatus
import org.springframework.web.bind.annotation.DeleteMapping
import org.springframework.web.bind.annotation.GetMapping
import org.springframework.web.bind.annotation.PatchMapping
import org.springframework.web.bind.annotation.PathVariable
import org.springframework.web.bind.annotation.PostMapping
import org.springframework.web.bind.annotation.RequestBody
import org.springframework.web.bind.annotation.RequestMapping
import org.springframework.web.bind.annotation.RequestParam
import org.springframework.web.bind.annotation.RestController
import org.springframework.web.server.ResponseStatusException
import java.time.Duration
import java.time.Instant
import java.time.LocalDate
import java.time.ZoneOffset
import kotlin.math.min

@RestController
@RequestMapping("/api/movies")
class MoviesController(
    private val repository: MovieQueryRepository,
    private val existingMovieWatchCreateFlowService: ExistingMovieWatchCreateFlowService,
    private val movieWatchRemovalService: MovieWatchRemovalService,
    private val movieTermsService: MovieTermsService,
    private val movieCreditsService: MovieCreditsService,
    private val movieListsService: MovieListsService,
) {
    @GetMapping("/library")
    fun library(
        @RequestParam(defaultValue = "20") limit: Int,
        @RequestParam(required = false) cursor: String?,
        @RequestParam(defaultValue = "false") unwatched: Boolean,
    ): MoviesLibraryResponse = repository.library(limit, cursor, unwatched)

    @GetMapping("/recent")
    fun recent(
        @RequestParam(defaultValue = "20") limit: Int,
        @RequestParam(required = false) cursor: String?,
    ): MoviesRecentResponse = repository.recent(limit, cursor)

    @GetMapping("/{movieId}")
    fun details(
        @PathVariable movieId: Long,
    ): MovieDetailsResponse = repository.getMovieDetails(movieId)

    @GetMapping("/slug/{slug}")
    fun detailsBySlug(
        @PathVariable slug: String,
    ): MovieDetailsResponse = repository.getMovieDetailsBySlug(slug)

    @GetMapping("/companies/{slug}")
    fun companyDetails(
        @PathVariable slug: String,
    ): MovieCompanyDetailsResponse = repository.getMovieCompanyDetails(slug)

    @GetMapping("/lists")
    fun lists(): List<MovieListSummaryDto> = movieListsService.listAll()

    @GetMapping("/lists/{slug}")
    fun listDetails(
        @PathVariable slug: String,
    ): MovieListDetailsResponse = repository.getMovieListDetails(slug)

    @GetMapping("/terms/{kind}/{slug}")
    fun termDetails(
        @PathVariable kind: String,
        @PathVariable slug: String,
    ): MovieTermDetailsResponse {
        val normalizedKind = kind.trim().uppercase()
        if (normalizedKind != "GENRE" && normalizedKind != "TAG") {
            throw ResponseStatusException(HttpStatus.BAD_REQUEST, "kind inválido")
        }
        return repository.getMovieTermDetails(normalizedKind, slug)
    }

    @GetMapping("/terms/search")
    fun searchTerms(
        @RequestParam q: String,
        @RequestParam kind: String,
        @RequestParam(defaultValue = "8") limit: Int,
    ): List<MovieTermSuggestionDto> {
        val normalizedKind = kind.trim().uppercase()
        if (normalizedKind != "GENRE" && normalizedKind != "TAG") {
            throw ResponseStatusException(HttpStatus.BAD_REQUEST, "kind inválido")
        }
        return repository.searchMovieTerms(q, normalizedKind, normalizeLimit("limit", limit))
    }

    @PostMapping("/{movieId}/credits/tmdb-candidates")
    fun movieTmdbCreditCandidates(
        @PathVariable movieId: Long,
    ): MovieTmdbCreditCandidatesResponse = movieCreditsService.fetchTmdbCandidates(movieId)

    @PostMapping("/{movieId}/credits/from-tmdb")
    fun importMovieTmdbCredit(
        @PathVariable movieId: Long,
        @RequestBody request: MovieTmdbCreditImportRequest,
    ): PersonCreditDto = movieCreditsService.importTmdbCredit(movieId, request)

    @PostMapping("/{movieId}/people")
    fun linkExistingPerson(
        @PathVariable movieId: Long,
        @RequestBody request: PersonLinkRequest,
    ): PersonCreditDto = movieCreditsService.linkExistingPerson(movieId, request)

    @DeleteMapping("/{movieId}/people/{personId}")
    fun removePersonCredit(
        @PathVariable movieId: Long,
        @PathVariable personId: Long,
        @RequestParam category: String,
    ) = movieCreditsService.removeCredit(movieId, personId, category)

    @PostMapping("/lists")
    fun createList(
        @RequestBody request: MovieListCreateRequest,
    ): MovieListSummaryDto = movieListsService.create(request)

    @PostMapping("/{movieId}/lists")
    fun attachMovieToList(
        @PathVariable movieId: Long,
        @RequestBody request: MovieListAttachRequest,
    ): MovieListSummaryDto = movieListsService.attachMovie(movieId, request)

    @DeleteMapping("/{movieId}/lists/{listId}")
    fun removeMovieFromList(
        @PathVariable movieId: Long,
        @PathVariable listId: Long,
    ) {
        movieListsService.removeMovie(movieId, listId)
    }

    @PostMapping("/lists/{listId}/order")
    fun updateMovieListOrder(
        @PathVariable listId: Long,
        @RequestBody request: MovieListOrderUpdateRequest,
    ) {
        movieListsService.updateOrder(listId, request)
    }

    @PatchMapping("/lists/{listId}/cover")
    fun updateMovieListCover(
        @PathVariable listId: Long,
        @RequestBody request: MovieListCoverUpdateRequest,
    ): MovieListSummaryDto = movieListsService.updateCover(listId, request)

    @PostMapping("/{movieId}/terms")
    fun addTerm(
        @PathVariable movieId: Long,
        @RequestBody request: MovieTermCreateRequest,
    ): MovieTermDto = movieTermsService.addTerm(movieId, request)

    @PostMapping("/{movieId}/terms/{termId}/visibility")
    fun updateMovieTermVisibility(
        @PathVariable movieId: Long,
        @PathVariable termId: Long,
        @RequestBody request: MovieTermVisibilityRequest,
    ): MovieTermDto = movieTermsService.updateMovieVisibility(movieId, termId, request.hidden)

    @PostMapping("/terms/{termId}/visibility")
    fun updateGlobalTermVisibility(
        @PathVariable termId: Long,
        @RequestBody request: MovieTermVisibilityRequest,
    ): MovieTermDto = movieTermsService.updateGlobalVisibility(termId, request.hidden)

    @PostMapping("/{movieId}/watches")
    fun addWatch(
        @PathVariable movieId: Long,
        @RequestBody request: ExistingMovieWatchCreateRequest,
    ): ManualMovieWatchCreateResponse = existingMovieWatchCreateFlowService.execute(movieId, request.watchedAt)

    @DeleteMapping("/{movieId}/watches/{watchId}")
    fun deleteWatch(
        @PathVariable movieId: Long,
        @PathVariable watchId: Long,
    ) {
        movieWatchRemovalService.remove(movieId, watchId)
    }

    @GetMapping("/search")
    fun search(
        @RequestParam q: String,
        @RequestParam(defaultValue = "10") limit: Int,
    ): MoviesSearchResponse = repository.search(q, limit)

    @GetMapping("/summary")
    fun summary(
        @RequestParam(defaultValue = "month") range: String,
        @RequestParam(required = false) start: Instant?,
        @RequestParam(required = false) end: Instant?,
    ): MoviesSummaryResponse {
        val (resolvedStart, resolvedEnd) = resolveRange(range, start, end)
        return repository.summary(resolvedStart, resolvedEnd)
    }

    @GetMapping("/stats")
    fun stats(): MoviesStatsResponse = repository.stats()

    @GetMapping("/year/{year}")
    fun byYear(
        @PathVariable year: Int,
        @RequestParam(defaultValue = "200") limitWatched: Int,
        @RequestParam(defaultValue = "200") limitUnwatched: Int,
    ): MoviesByYearResponse {
        val validatedYear = validateYear(year)
        val resolvedLimitWatched = normalizeLimit("limitWatched", limitWatched)
        val resolvedLimitUnwatched = normalizeLimit("limitUnwatched", limitUnwatched)
        val (start, end) = yearRange(validatedYear)
        return repository.byYear(
            year = validatedYear,
            start = start,
            end = end,
            limitWatched = resolvedLimitWatched,
            limitUnwatched = resolvedLimitUnwatched,
        )
    }

    private fun resolveRange(
        range: String,
        start: Instant?,
        end: Instant?,
    ): Pair<Instant, Instant> =
        when (range.lowercase()) {
            "month" -> {
                val e = Instant.now()
                val s = e.minus(Duration.ofDays(30))
                s to e
            }
            "year" -> {
                val e = Instant.now()
                val s = e.minus(Duration.ofDays(365))
                s to e
            }
            "custom" -> {
                if (start == null || end == null) {
                    throw IllegalArgumentException("range inválido")
                }
                start to end
            }
            else -> throw IllegalArgumentException("range inválido")
        }

    private fun validateYear(year: Int): Int {
        val maxYear = LocalDate.now(ZoneOffset.UTC).year + 1
        if (year < 1900 || year > maxYear) {
            throw ResponseStatusException(HttpStatus.BAD_REQUEST, "year inválido")
        }
        return year
    }

    private fun normalizeLimit(
        name: String,
        value: Int,
    ): Int {
        if (value < 1) {
            throw ResponseStatusException(HttpStatus.BAD_REQUEST, "$name deve ser >= 1")
        }
        return min(value, 1000)
    }

    private fun yearRange(year: Int): Pair<Instant, Instant> {
        val start = LocalDate.of(year, 1, 1).atStartOfDay(ZoneOffset.UTC).toInstant()
        val end = LocalDate.of(year, 12, 31).atTime(23, 59, 59).toInstant(ZoneOffset.UTC)
        return start to end
    }
}